Create password_update.sh
Sometimes when changing the password, whether in the domain or locally, the system does not automatically synchronize the new password with keychain to secure tocken which is manifested by the need to use the old password to decrypt the disk and the new one to log in. The script automates the manual synchronization process using shell commands.